collaspe defaulthandlercontext into filterhandler
